+++ CHANGELOG.md
@@ -26,6 +26,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 (key *Key) GetEntity() *openpgp.Entity
 ```
 
+- `EncryptSessionKey` now creates encrypted key packet for each valid key in keyring. Return is byte array with concatenated key packets.
+
 ### Changed
 - Improved key and message armoring testing